+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.test.ts
@@ -1,3 +1,4 @@
+import { deriveStateFromMetadata } from '@metamask/base-controller';
 import type nock from 'nock';
 
 import { mockUserStorageMessenger } from './__fixtures__/mockMessenger';
@@ -752,5 +753,91 @@ describe('UserStorageController', () => {
       messengerMocks.baseMessenger.publish('KeyringController:unlock');
       expect(await controller.getStorageKey()).toBe(MOCK_STORAGE_KEY);
     });
+  });
+});
+
+describe('metadata', () => {
+  it('includes expected state in debug snapshots', () => {
+    const controller = new UserStorageController({
+      messenger: mockUserStorageMessenger().messenger,
+    });
+
+    expect(
+      deriveStateFromMetadata(
+        controller.state,
+        controller.metadata,
+        'anonymous',
+      ),
+    ).toMatchInlineSnapshot(`
+      Object {
+        "isAccountSyncingEnabled": true,
+        "isBackupAndSyncEnabled": true,
+        "isContactSyncingEnabled": true,
+      }
+    `);
+  });
+
+  it('includes expected state in state logs', () => {
+    const controller = new UserStorageController({
+      messenger: mockUserStorageMessenger().messenger,
+    });
+
+    expect(
+      deriveStateFromMetadata(
+        controller.state,
+        controller.metadata,
+        'includeInStateLogs',
+      ),
+    ).toMatchInlineSnapshot(`
+      Object {
+        "hasAccountSyncingSyncedAtLeastOnce": false,
+        "isAccountSyncingEnabled": true,
+        "isAccountSyncingReadyToBeDispatched": false,
+        "isBackupAndSyncEnabled": true,
+        "isContactSyncingEnabled": true,
+      }
+    `);
+  });
+
+  it('persists expected state', () => {
+    const controller = new UserStorageController({
+      messenger: mockUserStorageMessenger().messenger,
+    });
+
+    expect(
+      deriveStateFromMetadata(controller.state, controller.metadata, 'persist'),
+    ).toMatchInlineSnapshot(`
+      Object {
+        "hasAccountSyncingSyncedAtLeastOnce": false,
+        "isAccountSyncingEnabled": true,
+        "isAccountSyncingReadyToBeDispatched": false,
+        "isBackupAndSyncEnabled": true,
+        "isContactSyncingEnabled": true,
+      }
+    `);
+  });
+
+  it('exposes expected state to UI', () => {
+    const controller = new UserStorageController({
+      messenger: mockUserStorageMessenger().messenger,
+    });
+
+    expect(
+      deriveStateFromMetadata(
+        controller.state,
+        controller.metadata,
+        'usedInUi',
+      ),
+    ).toMatchInlineSnapshot(`
+      Object {
+        "hasAccountSyncingSyncedAtLeastOnce": false,
+        "isAccountSyncingEnabled": true,
+        "isAccountSyncingReadyToBeDispatched": false,
+        "isBackupAndSyncEnabled": true,
+        "isBackupAndSyncUpdateLoading": false,
+        "isContactSyncingEnabled": true,
+        "isContactSyncingInProgress": false,
+      }
+    `);
   });
 });