+++ packages/profile-sync-controller/src/controllers/user-storage/UserStorageController.ts
@@ -81,24 +81,34 @@ export const defaultState: UserStorageControllerState = {
 
 const metadata: StateMetadata<UserStorageControllerState> = {
   isBackupAndSyncEnabled: {
+    includeInStateLogs: true,
     persist: true,
     anonymous: true,
+    usedInUi: true,
   },
   isBackupAndSyncUpdateLoading: {
+    includeInStateLogs: false,
     persist: false,
     anonymous: false,
+    usedInUi: true,
   },
   isAccountSyncingEnabled: {
+    includeInStateLogs: true,
     persist: true,
     anonymous: true,
+    usedInUi: true,
   },
   isContactSyncingEnabled: {
+    includeInStateLogs: true,
     persist: true,
     anonymous: true,
+    usedInUi: true,
   },
   isContactSyncingInProgress: {
+    includeInStateLogs: false,
     persist: false,
     anonymous: false,
+    usedInUi: true,
   },
 };